+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.24.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.23.1...v1.24.0) (2022-07-14)
+
+
+### Features
+
+* APPS-1649 add footer primary and sock ([#54](https://github.com/UCLALibrary/ucla-library-website-components/issues/54)) ([9d427f8](https://github.com/UCLALibrary/ucla-library-website-components/commit/9d427f84ccbb34a7702e729e3ebde1cb4f9b9653))
+
+## [1.23.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.23.0...v1.23.1) (2022-07-14)
+
+
+### Bug Fixes
+
+* add media gallery stories ([#59](https://github.com/UCLALibrary/ucla-library-website-components/issues/59)) ([29d6461](https://github.com/UCLALibrary/ucla-library-website-components/commit/29d6461b6c5ccec7c802bd919528bd50e4c4ea26))
+
 # [1.23.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.22.1...v1.23.0) (2022-07-14)